Rename dafault_checksums folder to tracked_checksums
