+++ droidlet/tools/artifact_scripts/compute_checksum.py
@@ -1,7 +1,7 @@
 # Copyright (c) Facebook, Inc. and its affiliates.
 
 """
-This script computes hashes for the given local artifact directories and saves them to data_scripts/default_checksums/
+This script computes hashes for the given local artifact directories and saves them to artifact_scripts/tracked_checksums/
 """
 import os
 import subprocess
@@ -36,7 +36,7 @@ def compute_checksum_for_directory(
         model_name = "nlu"
 
     print("Now computing hashes ...")
-    compute_shasum_script_path = os.path.join(ROOTDIR, 'droidlet/tools/data_scripts/checksum_fn.sh')
+    compute_shasum_script_path = os.path.join(ROOTDIR, 'droidlet/tools/artifact_scripts/checksum_fn.sh')
     checksum_name = ''
     artifact_folder_name = ''
     if artifact_type == "models":
@@ -56,7 +56,7 @@ def compute_checksum_for_directory(
         checksum_name = 'datasets.txt'
 
     artifact_path = os.path.join(ROOTDIR, 'droidlet/artifacts', artifact_folder_name)
-    checksum_write_path = os.path.join(ROOTDIR, 'droidlet/tools/data_scripts/default_checksums/' + checksum_name)
+    checksum_write_path = os.path.join(ROOTDIR, 'droidlet/tools/artifact_scripts/tracked_checksums/' + checksum_name)
     result = subprocess.check_output([compute_shasum_script_path, artifact_path, checksum_write_path],
                                      text=True)
     print(result)
