+++ droidlet/tools/artifact_scripts/tracked_checksums/craftassist_perception.txt
