+++ droidlet/tools/artifact_scripts/tracked_checksums/locobot_perception.txt
