+++ tools/README.md
@@ -5,6 +5,5 @@ This folder contains tools for the project:
 - `annotation_tools` has a suite of annotation tools for each sub-component of the project.
 - `codetools` has scripts that maintain code styling and code quality.
 - `dashboard_web` has the frontend component and build files of the project dashboard.
-- `data_scripts` has the scripts that maintain fetching and uploading data to our S3 bucket.
 - `docker` has the docker setup for the project.
 - `nsp_scripts` has scripts used by the semantic parser and surrounding datasets.
